int vs. long fixes

+++ indigo_bus/indigo_base64.h
@@ -39,9 +39,9 @@
 extern "C" {
 #endif
 
-extern int base64_encode(unsigned char *out, const unsigned char *in, int inlen);
-extern int base64_decode_fast(unsigned char *out, const unsigned char *in, int inlen);
-extern int base64_decode_fast_nl(unsigned char *out, const unsigned char *in, int inlen);
+extern long base64_encode(unsigned char *out, const unsigned char *in, long inlen);
+extern long base64_decode_fast(unsigned char *out, const unsigned char *in, long inlen);
+extern long base64_decode_fast_nl(unsigned char *out, const unsigned char *in, long inlen);
 
 #ifdef __cplusplus
 }
